// src/interfaces/Options.d.ts
export interface SharedElementTransition {
    fromId: string;
    toId: string;
    duration?: number;
    interpolation?: Interpolation;
}
export interface ElementTransition {
    id: string;
    alpha?: AppearingElementAnimation | DisappearingElementAnimation;
    translationX?: AppearingElementAnimation | DisappearingElementAnimation;
    translationY?: AppearingElementAnimation | DisappearingElementAnimation;
    scaleX?: AppearingElementAnimation | DisappearingElementAnimation;
    scaleY?: AppearingElementAnimation | DisappearingElementAnimation;
    rotationX?: AppearingElementAnimation | DisappearingElementAnimation;
    rotationY?: AppearingElementAnimation | DisappearingElementAnimation;
    x?: AppearingElementAnimation | DisappearingElementAnimation;
    y?: AppearingElementAnimation | DisappearingElementAnimation;
}
// ...
export interface OptionsAnimationPropertyConfig {
    /**
     * Animate from this value, ex. 0
     */
    from?: number;
    /**
     * Animate to this value, ex. 1
     */
    to?: number;
    /**
     * Animation duration
     * @default 300
     */
    duration?: number;
    /**
     * Animation delay
     * @default 0
     */
    startDelay?: number;
    /**
     * Animation interplation
     */
    interpolation?: Interpolation;
}
/**
 * Used to animate the actual content added to the hierarchy.
 * Content can be a React component (component) or any other layout (Stack, BottomTabs etc)
 */
export interface ScreenAnimationOptions {
    /**
     * Animate the element over x value
     */
    x?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over y value
     */
    y?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over translateX
     */
    translationX?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over translateY
     */
    translationY?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over opacity
     */
    alpha?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over scaleX
     */
    scaleX?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over scaleY
     */
    scaleY?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over rotationX
     */
    rotationX?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over rotationY
     */
    rotationY?: OptionsAnimationPropertyConfig;
    /**
     * Animate the element over rotation
     */
    rotation?: OptionsAnimationPropertyConfig;
    /**
     * Wait for the root view to render before start animation
     */
    waitForRender?: boolean;
    /**
     * Enable or disable the animation
     * @default true
     */
    enabled?: boolean;
}
// ...
export interface ViewAnimationOptions extends ScreenAnimationOptions {
    /**
     * ID of the Top Bar we want to animate
     */
    id?: string;
}
export interface TopBarAnimationOptions extends ViewAnimationOptions {
    /**
     * Animation of the top-bar's background color, in case the top-bar background color
     * has been explicitly specified.
     *
     * Applicable only in transition of screens with color (non-component) backgrounds.
     */
    bkgColor?: ColorAnimationOptions;
}
export interface StatusBarAnimationOptions extends ViewAnimationOptions {
    /**
     * Animation of the status-bar's background color, in case its background color
     * has been explicitly specified.
     */
    bkgColor?: ColorAnimationOptions;
}
export interface EnterExitAnimationOptions {
    /**
     * Animate opening component
     */
    enter?: ViewAnimationOptions;
    /**
     * Animate closing component
     */
    exit?: ViewAnimationOptions;
}
export interface OldModalAnimationOptions extends ViewAnimationOptions {
    /**
     * Animations to be applied on elements which are shared between the appearing and disappearing screens
     */
    sharedElementTransitions?: SharedElementTransition[];
    /**
     * Animations to be applied on views in the appearing or disappearing screens
     */
    elementTransitions?: ElementTransition[];
}
export interface ModalAnimationOptions {
    /**
     * Animate opening modal
     */
    enter?: ViewAnimationOptions;
    /**
     * Animate closing modal
     */
    exit?: ViewAnimationOptions;
    /**
     * Animations to be applied on elements which are shared between the appearing and disappearing screens
     */
    sharedElementTransitions?: SharedElementTransition[];
    /**
     * Animations to be applied on views in the appearing or disappearing screens
     */
    elementTransitions?: ElementTransition[];
}
/**
 * Used for describing stack commands animations.
 */
export interface StackAnimationOptions {
    /**
     * Wait for the View to render before start animation
     */
    waitForRender?: boolean;
    /**
     * Enable or disable the animation
     * @default true
     */
    enabled?: boolean;
    /**
     * Configure animations for the top bar
     */
    topBar?: TopBarAnimationOptions | {
        enter?: TopBarAnimationOptions;
        exit?: TopBarAnimationOptions;
    };
    /**
     * Configure animations for the status bar (typically aligned
     * with the top-bar's)
     */
    statusBar?: StatusBarAnimationOptions | {
        enter?: StatusBarAnimationOptions;
        exit?: StatusBarAnimationOptions;
    };
    /**
     * Configure animations for the bottom tabs
     */
    bottomTabs?: ViewAnimationOptions | {
        enter?: ViewAnimationOptions;
        exit?: ViewAnimationOptions;
    };
    /**
     * Configure animations for the content (Screen)
     */
    content?: ViewAnimationOptions | {
        enter?: ViewAnimationOptions;
        exit?: ViewAnimationOptions;
    };
    /**
     * Animations to be applied on elements which are shared between the appearing and disappearing screens
     */
    sharedElementTransitions?: SharedElementTransition[];
    /**
     * Animations to be applied on views in the appearing or disappearing screens
     */
    elementTransitions?: ElementTransition[];
}
/**
 * Used for configuring command animations
 */
export interface AnimationOptions {
    /**
     * Configure the setStackRoot animation
     */
    setStackRoot?: ViewAnimationOptions | StackAnimationOptions;
    /**
     * Configure the setRoot animation
     */
    setRoot?: ViewAnimationOptions | EnterExitAnimationOptions;
    /**
     * Configure the animation of the pushed screen
     * #### (Android specific)
     */
    push?: StackAnimationOptions;
    /**
     * Configure what animates when a screen is popped
     */
    pop?: StackAnimationOptions;
    /**
     * Configure what animates when modal is shown
     */
    showModal?: OldModalAnimationOptions | ModalAnimationOptions;
    /**
     * Configure what animates when modal is dismissed
     */
    dismissModal?: OldModalAnimationOptions | ModalAnimationOptions;
}